% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_archive.R
\name{dp_archive}
\alias{dp_archive}
\title{Archive data product}
\usage{
dp_archive(project_path = ".", ...)
}
\arguments{
\item{project_path}{path to project}

\item{...}{other parameters e.g. verbose = T}
}
\value{
TRUE
}
\description{
This undoes dp_deploy. It removes the deployed data product and
updates the dpboard-log
}
