% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_archive.R
\name{dp_archiveCore}
\alias{dp_archiveCore}
\title{The dp_archive is a wrapper around this.
Reason: With S3 generic methods, function calls as defaults parameters are
not recognized as the class of the object they return}
\usage{
dp_archiveCore(conf, dp_name, pin_version, git_info, ...)
}
\description{
The dp_archive is a wrapper around this.
Reason: With S3 generic methods, function calls as defaults parameters are
not recognized as the class of the object they return
}
\keyword{internal}
