% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_deploy.R
\name{dp_deploy}
\alias{dp_deploy}
\title{Deploy data product}
\usage{
dp_deploy(project_path = ".", ...)
}
\arguments{
\item{project_path}{path to project}

\item{...}{other parameters e.g. verbose = T}
}
\value{
TRUE
}
\description{
Deploys a data product to the where the board is set (e.g.
AWS s3)
}
