% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_deploy.R
\name{dp_deployCore}
\alias{dp_deployCore}
\title{The dp_deploy is a wrapper around this.
Reason: With S3 generic methods, function calls as defaults parameters are
not recognized as the class of the object they return}
\usage{
dp_deployCore(conf, project_path, d, dlog, git_info, type, ...)
}
\description{
The dp_deploy is a wrapper around this.
Reason: With S3 generic methods, function calls as defaults parameters are
not recognized as the class of the object they return
}
\keyword{internal}
