% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_sync.R
\name{dpinput_sync}
\alias{dpinput_sync}
\title{Sync Input Data to Remote}
\usage{
dpinput_sync(conf, input_map, verbose = F, ...)
}
\arguments{
\item{conf}{environment containing all config.R variables. See \code{dpconf_get()}}

\item{input_map}{object containing all input data to be synced. See \code{map_input()}}

\item{verbose}{T/F}

\item{...}{other parameters e.g. verbose = T}
}
\value{
synced_map this is input_map with sync status added to metadata
}
\description{
Sync input data for a data product build to a remote like a
labkey or AWS S3
}
