% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gitinfo_validate}
\alias{gitinfo_validate}
\title{Validate git info for deploy}
\usage{
gitinfo_validate(project_path, verbose = F)
}
\arguments{
\item{project_path}{path to project}

\item{verbose}{F if TRUE prints process details}
}
\value{
git_info, a list containing git information
}
\description{
Validates and extracts gitinfo per deploy requirements
}
\keyword{internal}
