% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpdeploy-package.R
\docType{package}
\name{dpdeploy-package}
\alias{dpdeploy}
\alias{dpdeploy-package}
\title{dpdeploy: A package to manage deployment of data products}
\description{
An R package to manage deployment of data product to remote.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/amashadihossein/dpdeploy/issues}
}

}
\author{
\strong{Maintainer}: Afshin Mashadi-Hossein \email{amashadihossein@gmail.com}

Authors:
\itemize{
  \item Yue Jiang \email{rivehill@gmail.com}
}

}
\keyword{internal}
