% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dlog}
\alias{get_dlog}
\title{Get dlog}
\usage{
get_dlog(project_path)
}
\value{
dlog
}
\description{
Reads and format daap_log.yml making pasting values is key:value
pairs with delimmitter " > "
}
\keyword{internal}
