% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{is_valid_dp_repository}
\alias{is_valid_dp_repository}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dpbuild}{\code{\link[dpbuild]{is_valid_dp_repository}}}
}}

